+++ packages/contracts-rfq/test/FastBridgeV2.GasBench.Dst.ArbitraryCall.t.sol
@@ -4,13 +4,16 @@ pragma solidity ^0.8.20;
 import {FastBridgeV2GasBenchmarkDstTest} from "./FastBridgeV2.GasBench.Dst.t.sol";
 import {RecipientMock} from "./mocks/RecipientMock.sol";
 
-// solhint-disable func-name-mixedcase, ordering
+// solhint-disable func-name-mixedcase, no-empty-blocks
 contract FastBridgeV2GasBenchmarkDstArbitraryCallTest is FastBridgeV2GasBenchmarkDstTest {
     // To get an idea about how much overhead the arbitrary call adds to the relaying process, we use a mock
     // recipient that has the hook function implemented as a no-op.
     // The mocked callParams are chosen to be similar to the real use cases:
     // - user address
     // - some kind of ID to decide what to do with the tokens next
+
+    /// @notice We include an empty "test" function so that this contract does not appear in the coverage report.
+    function testFastBridgeV2GasBenchmarkDstArbitraryCallTest() external {}
 
     function setUp() public virtual override {
         // In the inherited tests userB is always used as the recipient of the tokens.