+++ packages/contracts-rfq/test/FastBridgeV2.t.sol
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
+import {BridgeTransactionV2Lib} from "../contracts/libs/BridgeTransactionV2.sol";
+
 import {IFastBridge} from "../contracts/interfaces/IFastBridge.sol";
 
 // solhint-disable-next-line no-unused-import
@@ -44,11 +46,42 @@ abstract contract FastBridgeV2Test is Test, IFastBridgeV2Errors {
     IFastBridgeV2.BridgeParamsV2 internal tokenParamsV2;
     IFastBridgeV2.BridgeParamsV2 internal ethParamsV2;
 
+    bytes internal mockRequestV1;
+    bytes internal invalidRequestV2;
+    bytes internal mockRequestV3;
+
+    /// @notice We include an empty "test" function so that this contract does not appear in the coverage report.
+    function testFastBridgeV2Test() external {}
+
+    function createInvalidRequestV2(bytes memory requestV2) public pure returns (bytes memory result) {
+        // Copy everything but the last byte
+        result = new bytes(requestV2.length - 1);
+        for (uint256 i = 0; i < result.length; i++) {
+            result[i] = requestV2[i];
+        }
+    }
+
+    function createMockRequestV3(bytes memory requestV2) public pure returns (bytes memory result) {
+        result = new bytes(requestV2.length);
+        // Set the version to 3
+        result[0] = 0x00;
+        result[1] = 0x03;
+        // Copy the rest of the request
+        for (uint256 i = 2; i < result.length; i++) {
+            result[i] = requestV2[i];
+        }
+    }
+
     function setUp() public virtual {
         srcToken = new MockERC20("SrcToken", 6);
         dstToken = new MockERC20("DstToken", 6);
         createFixtures();
+        mockRequestV1 = abi.encode(extractV1(tokenTx));
+        // Invalid V2 request is formed before `createFixturesV2` to ensure it's not using callParams
+        invalidRequestV2 = createInvalidRequestV2(BridgeTransactionV2Lib.encodeV2(tokenTx));
         createFixturesV2();
+        // Mock V3 request is formed after `createFixturesV2` to ensure it's using callParams if needed
+        mockRequestV3 = createMockRequestV3(BridgeTransactionV2Lib.encodeV2(ethTx));
         fastBridge = deployFastBridge();
         configureFastBridge();
         mintTokens();
@@ -231,11 +264,21 @@ abstract contract FastBridgeV2Test is Test, IFastBridgeV2Errors {
     }
 
     function getTxId(IFastBridgeV2.BridgeTransactionV2 memory bridgeTx) public pure returns (bytes32) {
-        return keccak256(abi.encode(bridgeTx));
+        return keccak256(BridgeTransactionV2Lib.encodeV2(bridgeTx));
     }
 
     function expectUnauthorized(address caller, bytes32 role) public {
         vm.expectRevert(abi.encodeWithSelector(IAccessControl.AccessControlUnauthorizedAccount.selector, caller, role));
+    }
+
+    function expectRevertInvalidEncodedTx() public {
+        vm.expectRevert(BridgeTransactionV2Lib.BridgeTransactionV2__InvalidEncodedTx.selector);
+    }
+
+    function expectRevertUnsupportedVersion(uint16 version) public {
+        vm.expectRevert(
+            abi.encodeWithSelector(BridgeTransactionV2Lib.BridgeTransactionV2__UnsupportedVersion.selector, version)
+        );
     }
 
     function cheatCollectedProtocolFees(address token, uint256 amount) public {