+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.40.12](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.40.11...@synapsecns/synapse-interface@0.40.12) (2024-10-21)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.40.11](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.40.10...@synapsecns/synapse-interface@0.40.11) (2024-10-18)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.40.10](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.40.9...@synapsecns/synapse-interface@0.40.10) (2024-10-15)
+
+
+### Bug Fixes
+
+* **synapse-interface:** translation period ([#3297](https://github.com/synapsecns/sanguine/issues/3297)) ([fd75dbb](https://github.com/synapsecns/sanguine/commit/fd75dbbbb6fec8ad2c35d5a16efb3a6a4812231a))
+
+
+
+
+
+## [0.40.9](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.40.8...@synapsecns/synapse-interface@0.40.9) (2024-10-15)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.40.8](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.40.7...@synapsecns/synapse-interface@0.40.8) (2024-10-11)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface